+++ node/index.ts
@@ -1,4 +1,10 @@
-import { Cached, ClientsConfig, LRUCache, Service } from '@vtex/api'
+import {
+  Cached,
+  ClientsConfig,
+  LRUCache,
+  ParamsContext,
+  Service,
+} from '@vtex/api'
 
 import { Clients } from './clients'
 import { saveIOMessage } from './events/saveIOMessage'
@@ -22,6 +28,7 @@ import { State } from './typings/Colossus'
 const TIMEOUT_MS = 3000
 const TRANSLATION_CONCURRENCY = 5
 const TRANSLATION_RETRIES = 3
+const CONCURRENCY = 10
 
 const tenantCacheStorage = new LRUCache<string, Cached>({
   max: 3000,
@@ -31,6 +38,9 @@ const appsCacheStorage = new LRUCache<string, Cached>({
   max: 3000,
 })
 
+metrics.trackCache('tenant', tenantCacheStorage)
+metrics.trackCache('apps', appsCacheStorage)
+
 const clients: ClientsConfig<Clients> = {
   implementation: Clients,
   options: {
@@ -38,6 +48,7 @@ const clients: ClientsConfig<Clients> = {
       memoryCache: appsCacheStorage,
     },
     default: {
+      concurrency: CONCURRENCY,
       retries: 2,
       timeout: TIMEOUT_MS,
     },
@@ -53,7 +64,7 @@ const clients: ClientsConfig<Clients> = {
   },
 }
 
-export default new Service<Clients, State>({
+export default new Service<Clients, State, ParamsContext>({
   clients,
   events: {
     broadcasterBrand: [