+++ CHANGELOG.md
@@ -6,6 +6,10 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Fixed
+- Upgrades to node 6.x
+- Improve clients configurations
+- Use correct slugify
 
 ## [0.7.0] - 2020-02-11
 ### Added